fix(procfile): use native size for line metadata
The procfile parser stores words in a pfwords array indexed by size_t, but each parsed line stored its first word index and per-line word count in uint32_t fields. On 64-bit builds, an extreme input with more than UINT32_MAX words or line words could truncate ffline.first or wrap ffline.words while the backing word array continues to use size_t.

Current vs potential bug: normal /proc and /sys inputs are far below UINT32_MAX words, so this is a potential overflow/truncation bug rather than a currently exercised production path. The type mismatch still made the parser metadata narrower than the storage it indexes.

Solution: store ffline.words and ffline.first as size_t and make procfile_lines_add() return a size_t pointer, so the parser increments the same-width metadata it later exposes through the existing procfile macros.

Functional impact: normal parsing behavior, macro names, bounds checks, and callers are unchanged. Only extreme inputs that previously truncated line metadata now retain the full native-size count and first-word index.

Performance impact: no new parser branches or per-character work. On 64-bit builds each ffline grows from 8 to 16 bytes, increasing line-metadata memory only in proportion to parsed line count; 32-bit builds keep the same field width.

Validation: inspected the staged diff and kept the change limited to procfile line metadata. Worker validation passed whitespace checks, a focused low-feature libnetdata build, and direct proc_interrupts syntax checking. Required local reviewer quorum approved the source-level fix.
(cherry picked from commit a15b62bd15cedabe7a32ac5a3be98760b5ecdafc)
(cherry picked from commit 2e394ceb4e167a7900b20e5780b35d2e2557b3bd)

+++ src/libnetdata/procfile/procfile.c
@@ -103,7 +103,7 @@ static inline void procfile_words_free(pfwords *fw) {
 // An array of lines
 
 NEVERNULL
-static inline uint32_t *procfile_lines_add(procfile *ff) {
+static inline size_t *procfile_lines_add(procfile *ff) {
     // netdata_log_debug(D_PROCFILE, PF_PREFIX ":   adding line %d at word %d", fl->len, first_word);
 
     pflines *fl = ff->lines;
@@ -181,7 +181,7 @@ static void procfile_parser(procfile *ff) {
     char quote = 0;                     // the quote character - only when in quoted string
     size_t opened = 0;                  // counts the number of open parenthesis
 
-    uint32_t *line_words = procfile_lines_add(ff);
+    size_t *line_words = procfile_lines_add(ff);
 
     while(s < e) {
         PF_CHAR_TYPE ct = separators[(unsigned char)(*s)];
